refactor: extract useDirectLink hook into its own file
